pyproject.toml
[project]
name = "copilot-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mario Taddeucci", email = "mariotaddeucci@gmx.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "github-copilot-sdk>=0.1.23",
    "gitpython>=3.1.46",
    "injector>=0.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-ulid>=3.1.0",
    "textual>=7.5.0",
]

[project.scripts]
copilot-team = "copilot_team:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pyrefly>=0.52.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
lint = "ruff check . && ruff format --check . && pyrefly check"
fix = "ruff check --fix . && ruff format ."
test = "pytest"

[tool.pyrefly]
search_path = ["src", "."]

